Splitting the requires_unified_shared_memory tests into multiple tests, using different combinations of default mapping, is_device_ptr and explicit mapping of pointers
